refactor(slip): extract addOdd into shared actions
