+++ src/features/offer/components/center-section/odd.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { selectOddById } from "../../state/selectors";
-import { addOdd } from "../../../slip";
+import { addOdd } from "../../../shared";
 
 export const Odd = ({ oddId }) => {
   const dispatch = useDispatch();
