+++ src/features/shared/index.js
@@ -1,0 +1,1 @@
+export { combineIds, getCurrentSlip, addOdd } from "./actions";
