+++ src/features/slip/index.js
@@ -1,3 +1,2 @@
 export { Slip } from "./components/slip";
 export { slipReducer } from "./state/slipSlice";
-export { addOdd } from "./state/actions";
